Add ActivityInstance and Generator for Activities
Those two new concepts are introducing the ability to run an activity _n_ times
based on a value that is within a context. It allows more complex operations such
as: running one activity for a date within a date range. Retry mechanisms are
also a lot more granular.

The ActivityInstance represents an execution of the same activity with a different
context. Each execution of an activity will have at least one instance, and if you
use a generator, it will have n instances.

Adding generators does not change the dependency tree. If you have an activity
that depends on another one, the dependency will wait until all instances have
completed.

@rantonmattei