Add the runner param.
One of the recent request was to move the runner out, so you could write:
`create(..., runner=Sync(), tasks=[list of tasks])`

Still unsure if we should move forward with this implementation. The runner object will
always have a strong dependency on the tasks it runs. Another alternative would be to
have a `run` method instead that takes a runner and a list of tasks.