Add heartbeat.
Heartbeat is set by default by AWS (to 600 seconds, which is
10 minutes. If an activity last longer than 10 minutes but
does not issue a heartbeat - SWF considers this activity has
failed.

The heartbeat mechanism works on our end at task level. We
gather all the different heartbeats (if one activity misses
the heartbeat, it is set to the SWF default.) – and figure
out which one is the longuest. We use this one as a
reference for our task.

Additional notes:
* Everytime an activity spawns a new task, the heartbeat is
triggered.
* Everytime an activity finishes a task, a heartbeat is
triggered (this is for the async task runner.)

@rantonmattei