Introducing Task Lists
Activity generators have been very useful: they allow to spawn n instances of
an activity based on values in the context. The same can be done for tasks (
which makes it very handy when you write the logic units of your activity.)

**Goal**:

The goal of the generators is to provide an wasy way to write a list of tasks
if transformations need to be done within the context. The generator should also
contribute to the calculation of the different timeout values.

An additional goal was to make sure it is backward compatible: the way flows
have been written until today will remain the same.

**Example**:

```python
@task.list
def create_client(context):
    yield create_user.fill(
        username='context.username',
        email='context.email')
    if context.get('context.credit_card'):
        yield create_credit_card.fill(
            username='context.username',
            credit_card='context.credit_card')
    yield send_email.fill(email='context.email')
```

@rantonmattei