Add Param.
Not all the values that needs to be passed to the tasks are from the context, they could be values
that are part of the flow configuration, and values that should not need to be sent to AWS (for
instance: database credentials.)

This features allows to parametrize the fill. You can use: StaticParam or Param. Param is the
default behavior, and StaticParam is a value passed to the activity without sending it to SWF.
Note: this is not compatible with external activities, external activities can only consume
values that have been sent to SWF.

Retrocompatibility is ensured: all value strings are turned into Param.

Example:

    print_country_id.fill(
        country_id=param.Param('generator.country_id'),
        static=param.StaticParam('Hello World'))

@rantonmattei